Property binding
Learning about binding something that the user could change (inputs), rather than use sugar syntax which is an easier way of doing something that looks cleaner.  Important thing to note is that in order for the binding to work, it requires the square brackets on the property name. This in turn indicates one way data flow syntax. Data comes from the class, binds to the template and looks at the property and sets the property content. Adding image file within "assets" folder as for some reason putting the image outside of the "assets" folders causes the image to be broken.