Event binding
Using multiple types of events to listen to. Blur event will be called after the user enters input and clicks out of the input box. Input event is called as the user is typing a value inside the input box. Click event is called when the user clicks the button element. These events will cause one way data flow to happen by running the function to the data binding which will receive an update from the component class down into the template