Two-way Binding
Importing FormsModule to allow us to perform two-way data binding. Changing value binding from [value] to [ngModel]. Reason being for this is that in the browser, ngModel will talk to the input itself automatically. Getting rid of handleInput and handleBlur functions for handleChange because "event.target" is not really a great way of doing things. The way this will work now is that we're property binding using [ngModel], listening to a change event as the (ngModelChange) is an event binding. So when something changes, handleChange function will get called and the event will get given to us (pass us an event string via ngModelChange). The way two-way data binding is implemented is in the form of [(ngModel)] which seems like a property binding with an event binding inside that we’re listening to. So in the browser there will be two input boxes, the first input box will be using the handleChange function which is one-way data binding, and the other input box will be using the two-way data binding syntax. So by using the square bracket notation, we're combining the two whilst making the syntax even shorter for us. Ideally we would use this for input only or textarea rather than components. Mainly because it's easier to manage the data and the state of a component and template by using one-way data flow and event binding.