Template #ref variable
Getting rid of FormsModule as we no longer need the ngModel and ngModelChange. What we’re doing here is when clicking the button, we’re passing in a value from the input using the template ref. The template ref variable #username, allows us to create a reference to a particular DOM node which is accessible anywhere in the template. So by passing the username.value into the handleClick function, we can log that new value in the console every time the button is clicked. The template ref gives us access to the particular DOM element, which essentially exports the properties so that we can access it anywhere in the template using that hash variable.