ngIf, * syntax and <ng-template>
Using ngIf Angular Directive on the div that has been initialised with an empty string, to display when the user has typed a value length that is greater than 2 (template expression). This is also referred to as a structural directive because we can add directives to particular components or templates. The purpose to the * is to not need the use of ng-template element which is a web component specification. Angular sits on top of the web platform which means that we can use things like the Shadow DOM, to allow us to create a particular template that Angular can then use to generate content from. Such as the div being placed inside the ng-template which takes away the use of the *.  The ng-template is a web component piece of HTML that could be used as a template elsewhere, or it can be used multiple times. The * is simply sugar syntax for not having to type the word ng-template. But with ng-template you have to use square bracket notation as we’re using property binding to evaluate the ngIf expression which still sits in the template itself. Which is also why you are unable to see it but the div in the DOM if you were to inspect element it.