ngFor and iterating collections
Using ngFor to iterate through an array of passengers that has a type of Passanger array. The interface of Passenger is needed because it acts as a descriptor of the virtual object. Also safeguards the data set so that if something that was previously a boolean becomes a string, an error would display indicating that something has an incorrect type. Or if a property that hasn't been defined in the interface is added into the list, it would error about how it doesn't exist in the interface. The passengers is rendered in the list element by first creating a local variable called passenger. The "of passengers" indicates a "for of loop" Javascript method which essentially gives us access to each individual passenger. The "let i = index" is used to display the particular index of a passenger in numerical order. The ng-template is there to highlight what the *ngFor syntax is doing under the hood.