ngClass and className bindings
Adding dynamic styling based on whether a passenger is checked in or checked out. Using a span element with a class called "status" to base the styling on.  Added the appropriate styling with the "&.checked-in" indicating that it is a chain of classes on the same element. Showcasing 2 ways of dynamic styling based on use case. Firstly the easiest would be by property binding which is going to access the class property. So what this is doing is adding a class of "checked-in" if the user is checked in via "passenger.checkedIn" being true. This approach is best if we only want to add a single class name.
Second being the ngClass approach which passes in an object, with the first property being "checked-in" with the key of "passenger.checkedIn". The other property "checked-out" is highlighting how the ngClass approach allows for multiple classes to be used on a particular element. Although there is no styling for "checked-out", it can be seen via inspecting the particular element.