ngStyle and style bindings
Showcasing two other approaches via style binding and ngStyle. With the style binding method it binds the backgroundColor property, which passes in an expression based on whether the passenger has checked in or not. This uses a ternary operator where if the passenger is checked in, it'll set the backgroundColor to green, else set it to red. This in turn dynamically adds a style to an element.
The ngStyle approach uses the backgroundColor as a key on the object, where it evaluates the expression and runs the ternary statement.
These four different approaches show you how to achieve the same results but whatever apporach is used depends on the requirements