Pipes for data transformation
Pipes can be seen as a function/data transformation mechanism that returns something new/do some kind of transformation on a data set or single value. Added checkInDate to the interface with a null check. But if there was a case that the API doesn't return null dates, then the interface would have to be updated to "checkInDate?: number", which means that it is an optional property. The p tag uses the json pipe on the passenger object so that when viewing it on the browser, instead of "object object", it converts the JavaScript object into readable JSON. Within the div to display the passengers check in date, we are using the date pipe to specify a particular date format. Also used a ternary expression to display the date if the checkInDate is available, if not then display "Not checked in". Added the uppercase pipe to showcase how pipes can be chained together. Just have to be wary that if uppercase pipe was before the date pipe, then an error would occur as the uppercase would not be able to perform the uppercasing on numbers.