Safe navigation operator
Created another interface of Child that contains name and age. Which will then be added to the Passenger interface as an array of Child objects. Inside the div tag will demonstrate the use of a safe navigation operator "passenger.children?.length" where if a child does exist, display the length of the children array, if not then don't display anything. The "|| 0" is used so that if the child doesn't exist, give it a value of 0.