Feature modules with @NgModule
Created a custom NG module. This is so that when we build new features called feature modules, we don't want everything to live in the root module. Reason being is that we'll avoid having tons of imports which would make it hard to use different modules across different applications. The idea is that all of the logic for a passenger dashboard should live inside a module which allows us to take the module and paste it into another application, which should work without the need to doing anything else.