Creating a container (smart) component
Moved passengers dashboard logic into a smart/container component called "passenger-dashboard.component" within the "passenger-dashbaord" folder. This is so that the logic no longer lives in root component and is called out via the <passenger-dashboard> selector. Moved the interfaces into it's own file called "passenger.interface" within the "models" folder. This allows us to break up the interfaces and act as a particular model so that we can type check against various files without exporting them and including random files. This also helps when introducing something like a service which will then provide us some data, will be able to import the passenger module and tell the service that we're expecting maybe a single passenger or an array of passengers from a HTTP response. Exported the dashboard module and imported it into the root module.