ngOnInit lifecycle hook
Learning about OnInit lifecycle hook where lifecycle hook is a function that gets called by Angular when something happens. So the OnInit lifecycle hook is when the component is initialised. The purpose of the OnInit is to implement dynamic data and logic such as data fetching inside it so that the component is actually ready. When the component is initialised, this is when the passengers then get bound to the component.