Presentational (dumb) components
Setting up stateless child components for when we pass data into those components. Adding the components to the declarations of the NgModule so that it tells Angular that these are valid components. No need to add them to the exports as we only export the passenger dashboard component because in the root module we're only using the dashbaord component so the Angular module only needs to know about this single component. If we were to declare one of the child components in the app.component file (add <passenger-count> into template) we actually don't need to export that child component. The count component will be used to input data into it to render some data about the airline passengers. For the data about how many passengers there are and how many are checked in will be part of the detail component to render these lists out as individual components.