Passing data into components with @Input
Displaying the passenger count by one way binding the array of passengers to the custom component. Count component now expects an input binding called items which is declared as a property which is going to be an array of passengers. Input decorator tells Angular that items is going to be an input binding. Created function to that first checks if there are no items available which returns 0. The function calculates how many passengers are checked in by returning a filtered amount of passengers from the passenger object of type Passenger, and the length of those that are checked in.